+++ packages/taro-cli/src/util/resolve_npm_files.js
@@ -121,6 +121,7 @@ function parseAst (ast, filePath, files, isProduction, npmConfig, buildAdapter =
                   let tempPathWithIndexJS = `${realRequirePath}${path.sep}index.js`
                   if (fs.existsSync(tempPathWithJS)) {
                     realRequirePath = tempPathWithJS
+                    requirePath += '.js'
                   } else if (fs.existsSync(tempPathWithIndexJS)) {
                     realRequirePath = tempPathWithIndexJS
                     requirePath += '/index.js'
@@ -154,7 +155,7 @@ function recursiveRequire (filePath, files, isProduction, npmConfig = {}, buildA
   if (REG_STYLE.test(path.basename(filePath))) {
     return
   }
-  fileContent = npmCodeHack(filePath, fileContent)
+  fileContent = npmCodeHack(filePath, fileContent, buildAdapter)
   try {
     const transformResult = wxTransformer({
       code: fileContent,
@@ -197,13 +198,17 @@ function recursiveRequire (filePath, files, isProduction, npmConfig = {}, buildA
   }
 }
 
-function npmCodeHack (filePath, content) {
+function npmCodeHack (filePath, content, buildAdapter) {
   const basename = path.basename(filePath)
   switch (basename) {
     case 'lodash.js':
     case '_global.js':
     case 'lodash.min.js':
-      content = content.replace(/Function\([\'"]return this[\'"]\)\(\)/, 'this')
+      if (buildAdapter === BUILD_TYPES.ALIPAY) {
+        content = content.replace(/Function\([\'"]return this[\'"]\)\(\)/, '{}')
+      } else {
+        content = content.replace(/Function\([\'"]return this[\'"]\)\(\)/, 'this')
+      }
       break
     case '_html.js':
       content = 'module.exports = false;'