+++ packages/taro-rn/src/component.js
@@ -7,7 +7,7 @@ class Component extends React.Component {
   constructor (props, context) {
     super(props, context)
     if (props.navigation) {
-      this.$router = props.navigation.state.params || {}
+      this.$router = {params: props.navigation.state.params || {}}
     }
   }