+++ packages/taro-router-rn/src/RefreshProvider.js
@@ -10,25 +10,29 @@ class RefreshProvider extends React.Component {
   }
 
   render () {
-    const {enablePullDownRefresh} = this.props
-    return (
-      <ScrollView
-        style={{flex: 1}}
-        scrollEventThrottle={5}
-        alwaysBounceVertical={false}
-        onScroll={this.onScroll}
-        refreshControl={
-          enablePullDownRefresh
-            ? <RefreshControl
-              refreshing={this.state.refreshing}
-              onRefresh={this.handlePullDownRefresh}
-            />
-            : null
-        }
-      >
-        {this.props.children && this.props.children}
-      </ScrollView>
-    )
+    const {enablePullDownRefresh, disableScroll} = this.props
+    if (disableScroll) {
+      return this.props.children
+    } else {
+      return (
+        <ScrollView
+          style={{flex: 1}}
+          scrollEventThrottle={5}
+          alwaysBounceVertical={false}
+          onScroll={this.onScroll}
+          refreshControl={
+            enablePullDownRefresh
+              ? <RefreshControl
+                refreshing={this.state.refreshing}
+                onRefresh={this.handlePullDownRefresh}
+              />
+              : null
+          }
+        >
+          {this.props.children && this.props.children}
+        </ScrollView>
+      )
+    }
   }
 
   onScroll = (e) => {